buoiTH5

+++ QuanLyBanHang/Data/HoaDonChiTiet.cs
@@ -12,12 +12,13 @@ namespace QuanLyBanHang.Data
         public int ID { get; set; }
         public int HoaDonID { get; set; }
         public int SanPhamID { get; set; }
-        public int SoLuongBan { get; set; }
+        public short SoLuongBan { get; set; }
         public int DonGiaBan { get; set; }
 
         public virtual HoaDon HoaDon { get; set; } = null!;
         public virtual SanPham SanPham { get; set; } = null!;
     }
+
     [NotMapped]
     public class DanhSachHoaDon_ChiTiet
     {
@@ -25,7 +26,7 @@ namespace QuanLyBanHang.Data
         public int HoaDonID { get; set; }
         public int SanPhamID { get; set; }
         public string TenSanPham { get; set; }
-        public int SoLuongBan { get; set; }
+        public short SoLuongBan { get; set; }
         public int DonGiaBan { get; set; }
         public int ThanhTien { get; set; }
     }
